代码优化。修复大量警告

+++ FileBrowser/FileBrowser/FileSelectControler/Tools/HSDownloadManager/MD5/NSString+Hash.m
@@ -4,7 +4,7 @@
 //  Created by Tom Corwine on 5/30/12.
 //
 
-#import "NSString+hash.h"
+#import "NSString+Hash.h"
 
 #import <CommonCrypto/CommonDigest.h>
 #import <CommonCrypto/CommonHMAC.h>
@@ -76,10 +76,10 @@
 
 #pragma mark - Helpers
 
-- (NSString *)stringFromBytes:(unsigned char *)bytes length:(int)length
+- (NSString *)stringFromBytes:(unsigned char *)bytes length:(NSUInteger)length
 {
 	NSMutableString *mutableString = @"".mutableCopy;
-	for (int i = 0; i < length; i++)
+	for (NSUInteger i = 0; i < length; i++)
 		[mutableString appendFormat:@"%02x", bytes[i]];
 	return [NSString stringWithString:mutableString];
 }
